requirements_offline.txt
# ═══════════════════════════════════════════════════════════════
# GOEIC OFFLINE DEPENDENCIES
# ═══════════════════════════════════════════════════════════════

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2

# Security
bcrypt==4.2.0
python-multipart==0.0.12

# Database
weaviate-client==4.9.3

# Local AI Models
sentence-transformers==3.3.1
torch==2.10.0
transformers==4.46.3

# Ollama Client
httpx==0.26.0
aiohttp==3.11.7

# Voice (Local Whisper)
# openai-whisper==20240930
edge-tts==6.1.9

# Web Scraping & Processing
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# Text Processing
langchain-text-splitters==0.3.2

# Excel/Document Processing
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2

# Utilities
tqdm==4.67.1
Pillow==11.0.0

# Optional: For better Arabic support
arabic-reshaper==3.0.0
python-bidi==0.4.2
